+++ packages/ui/src/components/shared/unsplash/index.ts
@@ -1,2 +1,1 @@
-export type * from "./index.types";
-export * from "./unsplash-provider";
+export * from "./unsplash";